feat(bsp): update HAL and memory map for CLINT support

+++ fpga/sw/platform/bsp/hal/hal_timer.c
@@ -1,4 +1,5 @@
 #include "hal_timer.h"
+#include "math_ops.h"
 
 // ============================================================================
 // FUNÇÕES DE DELAY (BLOQUEANTE)
@@ -10,10 +11,10 @@ void hal_timer_delay_us(uint32_t us) {
     
     // Otimização Bare-Metal:
     // 1 us = 100 ciclos (@ 100MHz).
-    // Limite seguro: ~42 segundos de delay.
-    uint32_t cycles_to_wait = us * (SYSTEM_CLOCK_HZ / 1000000);
+    // Nota: Cuidado com overflow se us for muito grande (uint32_t max * 100).
+    uint64_t cycles_to_wait = (uint64_t)us * (SYSTEM_CLOCK_HZ / 1000000);
     
-    while ((hal_timer_get_cycles() - start) < (uint64_t)cycles_to_wait);
+    while ((hal_timer_get_cycles() - start) < cycles_to_wait);
 
 }
 
@@ -22,8 +23,8 @@ void hal_timer_delay_ms(uint32_t ms) {
     uint64_t start = hal_timer_get_cycles();
     
     // 1 ms = 100.000 ciclos (@ 100MHz).
-    uint32_t cycles_to_wait = ms * (SYSTEM_CLOCK_HZ / 1000);
+    uint64_t cycles_to_wait = (uint64_t)ms * (SYSTEM_CLOCK_HZ / 1000);
     
-    while ((hal_timer_get_cycles() - start) < (uint64_t)cycles_to_wait);
+    while ((hal_timer_get_cycles() - start) < cycles_to_wait);
     
 }
